+++ tests/workflows/cancelWorkflow.test.ts
@@ -1,10 +1,10 @@
 import _ from "lodash";
 import { describe, beforeAll, test, expect } from "@jest/globals";
 import { Client } from "@avaprotocol/sdk-js";
-import { getAddress, generateSignature, SaltGlobal } from "./utils";
-import { createFromTemplate } from "./templates";
+import { getAddress, generateSignature, SaltGlobal } from "../utils/utils";
+import { createFromTemplate } from "../utils/templates";
 import { WorkflowStatus } from "@avaprotocol/types";
-import { getConfig } from "./envalid";
+import { getConfig } from "../utils/envalid";
 
 // Get environment variables from envalid config
 const { avsEndpoint, walletPrivateKey, factoryAddress } = getConfig();